Alternate Dynamic Dictionary syntax and semantics
Change the Dynamic Dictionary syntax to be more like javascript
object literal syntax, and the Compiler to create javascript
objects instead of HashedCollections.

Now you can write

        #{
                click: [ :evt | console log: evt ].
                'background-color': 'green' }

instead of

        #{
                'click' -> [ :evt | console log: evt ].
                'background-color' -> 'green' }

or #{ a: 1 + 1 } instead of #{ 'a' -> (1 + 1) }

and get a js object instead of a HashedCollection 10-15 times faster:

original dynamic dictionary:
        [ 100000 timesRepeat: [ #{ 'a' -> 1 } ] ] timeToRun. " 1724 "
this implementation:
        [ 100000 timesRepeat: [ #{ a: 1 } ] ] timeToRun. " 112 "
inline javascript:
        [ 100000 timesRepeat: [ <{ a: 1 }> ] ] timeToRun. " 94 "
noop:
        [ 100000 timesRepeat: [ ] ] timeToRun. " 107 "

Pure javascript objects whitout any additional properties are better for interfacing other js libs.
For example try this in a Workspace with the original implementation:

        'body' asJQuery css: #{ 'background-color' -> 'red' }

It throws a MNU: [object HTMLBodyElement] does not understand #keys
while it works as expected with the new implementation.